move util functions out of game
